Move basis vector optimisation code to new module.
- reinstate optimise_initial_basis_vectors option
- move shared pytest fixtrue to conftest
