+++ algorithms/indexing/indexer.py
@@ -1134,81 +1134,6 @@ class indexer_base(object):
         raise NotImplementedError()
 
 
-def optimise_basis_vectors(reciprocal_lattice_points, vectors):
-    optimised = flex.vec3_double()
-    for vector in vectors:
-        minimised = basis_vector_minimser(reciprocal_lattice_points, vector)
-        optimised.append(tuple(minimised.x))
-    return optimised
-
-
-from scitbx import lbfgs
-
-# Optimise the initial basis vectors as per equation 11.4.3.4 of
-# Otwinowski et al, International Tables Vol. F, chapter 11.4 pp. 282-295
-class basis_vector_target(object):
-    def __init__(self, reciprocal_lattice_points):
-        self.reciprocal_lattice_points = reciprocal_lattice_points
-        self._xyz_parts = self.reciprocal_lattice_points.parts()
-
-    def compute_functional_and_gradients(self, vector):
-        assert len(vector) == 3
-        two_pi_S_dot_v = 2 * math.pi * self.reciprocal_lattice_points.dot(vector)
-        f = -flex.sum(flex.cos(two_pi_S_dot_v))
-        sin_part = flex.sin(two_pi_S_dot_v)
-        g = flex.double(
-            [flex.sum(2 * math.pi * self._xyz_parts[i] * sin_part) for i in range(3)]
-        )
-        return f, g
-
-
-class basis_vector_minimser(object):
-    def __init__(
-        self,
-        reciprocal_lattice_points,
-        vector,
-        lbfgs_termination_params=None,
-        lbfgs_core_params=lbfgs.core_parameters(m=20),
-    ):
-        self.reciprocal_lattice_points = reciprocal_lattice_points
-        if not isinstance(vector, flex.double):
-            self.x = flex.double(vector)
-        else:
-            self.x = vector.deep_copy()
-        self.n = len(self.x)
-        assert self.n == 3
-        self.target = basis_vector_target(self.reciprocal_lattice_points)
-        self.minimizer = lbfgs.run(
-            target_evaluator=self,
-            termination_params=lbfgs_termination_params,
-            core_params=lbfgs_core_params,
-        )
-        # print "number of iterations:", self.minimizer.iter()
-
-    def compute_functional_and_gradients(self):
-        f, g = self.target.compute_functional_and_gradients(tuple(self.x))
-        # g_fd = _gradient_fd(self.target, tuple(self.x))
-        # from libtbx.test_utils import approx_equal
-        # assert approx_equal(g, g_fd, eps=1e-3)
-        return f, g
-
-    def callback_after_step(self, minimizer):
-        # print tuple(self.x)
-        return
-
-
-def _gradient_fd(target, vector, eps=1e-6):
-    grads = []
-    for i in range(len(vector)):
-        v = list(vector)
-        v[i] -= eps
-        tm, _ = target.compute_functional_and_gradients(v)
-        v[i] += 2 * eps
-        tp, _ = target.compute_functional_and_gradients(v)
-        grads.append((tp - tm) / (2 * eps))
-    return grads
-
-
 def reject_weight_outliers_selection(reflections, sigma_cutoff=5):
     from scitbx.math import basic_statistics
